+++ packages/snap-preact/components/src/themes/pike/components/molecules/checkbox.ts
@@ -3,9 +3,6 @@ import type { CheckboxProps, CheckboxTemplatesLegalProps } from '../../../../com
 import { ThemeComponent } from '../../../../providers';
 import { custom } from '../../custom';
 
-// static variables
-const darkGray = custom.utils.darkenColor();
-
 // CSS in JS style script for the Checkbox component
 const checkboxStyleScript = (props: CheckboxProps) => {
 	// eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -22,8 +19,8 @@ const checkboxStyleScript = (props: CheckboxProps) => {
 		sharedStyles,
 		{
 			...custom.styles.box('', 0),
+			borderColor: props.color || custom.colors.gray02,
 			'&.ss__checkbox--active': {
-				borderColor: darkGray,
 				backgroundColor: custom.colors.white,
 			},
 			'&.ss__checkbox--disabled': {